requirements.txt
beautifulsoup4
spotipy
requests
python-dotenv
tenacity
